Streamlit/requirements.txt
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
gitdb
GitPython
h11
httpcore
httpx
idna
Jinja2
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
narwhals
numpy
ollama
packaging
pandas
pillow
pip
protobuf
pyarrow
pydantic
pydantic_core
pydeck
Pygments
python-dateutil
python-dotenv
pytz
referencing
requests
rich
rpds-py
six
smmap
sniffio
streamlit
tenacity
toml
tornado
typing_extensions
tzdata
urllib3
